fix(VLM): デコーダをfp32化して実機の誤キャプションを解消（fp16 logits反転）
実機ログで確定: encoder出力は有限でMacと値域一致(min-46/max32.5)、mask全1も正常。
残るはデコーダのfp16数値のみ。実機のfp16デコーダがMacと乖離し、近接トークンの
argmaxが反転(2トークン目 Mac133'The'→実機140'Trump')して全系列破綻、言語モデルの
地の文を吐いていた。ANE回避(CPU+GPU)でも実機GPUのfp16で同様に壊れた。

- convert_florence.py: デコーダを compute_precision=FLOAT32(encoderはfp16のまま)
- VLMRuntime.argmaxRow を fp32/fp16 両dtype対応に
- captionModelVersion 4 で誤キャプション破棄＆付け直し
- 診断ログ(encoder有限性・生成ID)は残置

デコーダ 184→367MB(VLM計626MB)。footprintは要監視。

+++ scripts/convert_florence.py
@@ -84,14 +84,16 @@ enc_ml.save(os.path.join(OUT, "VLMVision.mlpackage")); print("saved VLMVision")
 PAD = lm_cfg.pad_token_id
 dec_ids0 = torch.full((1, MAXLEN), PAD, dtype=torch.int32); dec_ids0[0, 0] = lm_cfg.decoder_start_token_id
 dec_traced = torch.jit.trace(dec_mod, (dec_ids0, enc_hidden, enc_mask), check_trace=False)
-# encoder_hidden/mask は VLMVision の出力（fp16）をそのまま繋ぐため fp16 入力にする
-# （fp32 にすると Swift 側で毎回キャスト copy が要る）。
+# ⚠️ デコーダは compute_precision=FLOAT32。fp16 だと**実機**でのみ Mac と logits が乖離し、
+#    近接トークンの argmax が反転して全写真同一の無関係テキスト(言語モデルの地の文)を吐く
+#    （Mac は全 compute unit で正常・encoder は fp16 でも一致・case-studies 参照）。encoder は fp16 のまま。
+# encoder_hidden/mask 入力は fp16（VLMVision 出力を直結）→ デコーダ内部で fp32 に upcast される。
 dec_ml = ct.convert(dec_traced,
     inputs=[ct.TensorType(name="decoder_input_ids", shape=(1, MAXLEN), dtype=np.int32),
             ct.TensorType(name="encoder_hidden", shape=enc_hidden.shape, dtype=np.float16),
             ct.TensorType(name="encoder_mask", shape=enc_mask.shape, dtype=np.float16)],
     outputs=[ct.TensorType(name="logits")],
-    compute_precision=ct.precision.FLOAT16, minimum_deployment_target=ct.target.iOS17)
+    compute_precision=ct.precision.FLOAT32, minimum_deployment_target=ct.target.iOS17)
 dec_ml.save(os.path.join(OUT, "VLMDecoder.mlpackage")); print("saved VLMDecoder")
 
 # ---- トークナイザ資産（vocab / merges）＋ config ----
